feat: Add plane offset configuration for translation gizmo
Introduces a new `planeOffset` configuration option for the translation gizmo. This allows users to adjust the distance of the plane handles from the origin. The `TranslationGizmo` component has been updated to utilize this new configuration value for calculating the position of the plane handles, replacing the hardcoded offset. A corresponding UI element has been added to the preferences modal for easy adjustment of this setting.